documentation: Give general documentation for the `core` crate
