+++ backend/_core/rust/Cargo.toml
@@ -6,6 +6,7 @@ authors = [
     "Chloe Ross <chloe@launchbadge.com>",
 ]
 edition = "2018"
+publish = false
 
 [dependencies]
 anyhow = "1.0.32"
@@ -27,7 +28,4 @@ sqlproxy = ["db"]
 local = ["listenfd"]
 release = []
 sandbox = []
-s3 = []
-jwk = []
 db = ["sqlx"]
-algolia = []
