+++ backend/_core/rust/src/http.rs
@@ -1,16 +1,19 @@
 use std::net::{SocketAddr, TcpListener};
 
-#[cfg(feature = "local")]
+/// Creates a [`TcpListener`] via `listenfd` if possible.
+///
+/// [`TcpListener`]: https://doc.rust-lang.org/stable/std/net/struct.TcpListener.html
 pub fn get_tcp_fd() -> Option<TcpListener> {
-    listenfd::ListenFd::from_env().take_tcp_listener(0).unwrap()
+    #[cfg(feature = "local")]
+    let fd = listenfd::ListenFd::from_env().take_tcp_listener(0).unwrap();
+
+    #[cfg(not(feature = "local"))]
+    let fd = None;
+
+    fd
 }
 
-#[cfg(not(feature = "local"))]
-pub fn get_tcp_fd() -> Option<TcpListener> {
-    // we don't have listenfd here.
-    None
-}
-
+/// Get the port to run the server on.
 pub fn get_addr(default: u16) -> SocketAddr {
     let port = std::env::var("PORT").map_or(default, |it| it.parse().unwrap_or(default));
