--- backend/_core/rust/src/settings/s3.rs
@@ -1,7 +1,0 @@
-pub struct S3Settings {
-    pub endpoint: String,
-    pub bucket: String,
-    pub use_client: bool,
-    pub access_key_id: String,
-    pub secret_access_key: String,
-}
